feat(mon-pix): move anonymize user usecase in src
